+++ src/main/java/com/winnow/bestchoice/controller/OAuthController.java
@@ -2,8 +2,6 @@ package com.winnow.bestchoice.controller;
 
 import com.winnow.bestchoice.dto.CreateAccessTokenRequest;
 import com.winnow.bestchoice.dto.CreateAccessTokenResponse;
-import com.winnow.bestchoice.exception.ErrorResponse;
-import com.winnow.bestchoice.model.response.TokenResponse;
 import com.winnow.bestchoice.service.TokenService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,10 +9,13 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
 @RequiredArgsConstructor
 @RestController
 @Slf4j
-public class TokenApiController {
+public class OAuthController {
     private final TokenService tokenService;
 
     @PostMapping("/api/token")
@@ -25,14 +26,12 @@ public class TokenApiController {
                 .body(new CreateAccessTokenResponse(newAccessToken));
     }
 
-    @GetMapping("/token")
-    public ResponseEntity<TokenResponse> token(@RequestParam("token") String token) {
-        return ResponseEntity.ok().body(new TokenResponse(token, "OK"));
+    @GetMapping("/login/{provider}")
+    public void login(@PathVariable String provider, HttpServletResponse response) throws IOException {
+        response.sendRedirect("http://www.winnow-bestchoice.com:8080/oauth2/authorization/" + provider);
     }
 
-    @GetMapping("/error")
-    public ResponseEntity<?> error(@RequestParam("error") String message) {
-        return ResponseEntity.ok().body(message);
-    }
+
+
 
 }
