+++ src/main/resources/application.yml
@@ -49,7 +49,7 @@ spring:
             clientSecret: ${KAKAO_CLIENT_SECRET}
             clientAuthenticationMethod: post
             authorizationGrantType: authorization_code
-            redirectUri: http://www.winnow-bestchoice.com:8080/oauth2/code/kakao
+            redirectUri: http://www.winnow-bestchoice.com:8080/code/kakao
             scope:
               - profile_nickname
               - account_email
@@ -86,7 +86,7 @@ jwt:
 
 # cors 설정
 cors:
-  allowed-origins: http://www.winnow-bestchoice.com:3000
+  allowed-origins: http://localhost:5173
   allowed-methods: GET,POST,PUT,DELETE,OPTIONS
   allowed-headers: '*'
   max-age: 3600
