--- src/main/resources/templates/login.html
@@ -1,13 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <title>로그인</title>
-</head>
-<body>
-<h1>안녕하세요 베스트 초이스 로그인창입니다^^</h1>
-<a href="http://www.winnow-bestchoice.com:8080/oauth2/authorization/google?redirect_uri=http://www.winnow-bestchoice.com:8080/oauth2/code/google">구글 로그인</a>
-<a href="http://www.winnow-bestchoice.com:8080/oauth2/authorization/naver?redirect_uri=http://www.winnow-bestchoice.com:8080/oauth2/code/naver">네이버 로그인</a>
-<a href="http://www.winnow-bestchoice.com:8080/oauth2/authorization/kakao?redirect_uri=http://www.winnow-bestchoice.com:8080/oauth2/code/kakao">카카오 로그인</a>
-</body>
-</html>
